requirements.txt
asgiref==3.3.0
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
gunicorn==20.0.4
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
requests==2.25.0